Moves schema test to the appropriate test class
